refactor: 2단계 - 로또(자동)
1. Lotto matchingCount 함수 추가, createWinning 함수 수정
2. LottoNumber value class 변경
3. LottoStatistics class 추가
4. LottoWinning lottoResult 함수 변경
