+++ src/main/kotlin/lotto/model/Lotto.kt
@@ -8,15 +8,18 @@ class Lotto(val numbers: Set<LottoNumber>) {
         }
     }
 
+    fun matchingCount(lotto: Lotto) = numbers.count {
+        it in lotto.numbers
+    }
+
     companion object {
         private const val LOTTO_SIZE = 6
-        private const val DELIMITER = ","
 
         fun create() = Lotto(
             (LottoNumber.LOTTO_NUMBER_MIN..LottoNumber.LOTTO_NUMBER_MAX).shuffled().take(LOTTO_SIZE).sorted()
                 .map { LottoNumber(it) }.toSet()
         )
 
-        fun createWinning(numbers: String) = Lotto(numbers.split(DELIMITER).map { LottoNumber(it.toInt()) }.toSet())
+        fun createWinning(numbers: List<Int>) = Lotto(numbers.map { LottoNumber(it) }.toSet())
     }
 }
