+++ src/test/kotlin/lotto/model/LottoWinningTest.kt
@@ -5,9 +5,9 @@ import io.kotest.matchers.shouldBe
 
 class LottoWinningTest : BehaviorSpec({
     Given("1,2,3,4,5,6 당첨번호") {
-        val lottoWinning = LottoWinning("1,2,3,4,5,6")
+        val lottoWinning = LottoWinning("1,2,3,4,5,6".toNumbers())
         When("나의 로또 번호 1,2,3,7,8,9") {
-            val lotto = Lotto.createWinning("1,2,3,7,8,9")
+            val lotto = Lotto.createWinning("1,2,3,7,8,9".toNumbers())
             Then("3개 일치 이다") {
                 val result = lottoWinning.lottoResult(listOf(lotto)).toList()
                 val prize = result[0].first
@@ -16,7 +16,7 @@ class LottoWinningTest : BehaviorSpec({
         }
 
         When("나의 로또 번호 1,2,3,4,8,9") {
-            val lotto = Lotto.createWinning("1,2,3,4,8,9")
+            val lotto = Lotto.createWinning("1,2,3,4,8,9".toNumbers())
             Then("4개 일치 이다") {
                 val result = lottoWinning.lottoResult(listOf(lotto)).toList()
                 val prize = result[0].first
@@ -25,7 +25,7 @@ class LottoWinningTest : BehaviorSpec({
         }
 
         When("나의 로또 번호 1,2,3,4,5,6") {
-            val lotto = Lotto.createWinning("1,2,3,4,5,6")
+            val lotto = Lotto.createWinning("1,2,3,4,5,6".toNumbers())
             Then("6개 일치 이다") {
                 val result = lottoWinning.lottoResult(listOf(lotto)).toList()
                 val prize = result[0].first
@@ -36,14 +36,14 @@ class LottoWinningTest : BehaviorSpec({
         When("로또 번호 14개의 중 3개 일치 1개") {
             val lottoList = mutableListOf<Lotto>().apply {
                 repeat(13) {
-                    add(Lotto.createWinning("11,12,13,14,15,16"))
+                    add(Lotto.createWinning("11,12,13,14,15,16".toNumbers()))
                 }
-                add(Lotto.createWinning("1,2,3,7,8,9"))
+                add(Lotto.createWinning("1,2,3,7,8,9".toNumbers()))
             }
             Then("구매금액 14000 수익률 0.35") {
                 val amount = 14000
                 val result = lottoWinning.lottoResult(lottoList)
-                val total = lottoWinning.totalStatistics(result, amount)
+                val total = LottoStatistics.totalStatistics(result, amount)
 
                 println(" total $total")
                 total.toDecimalPoint() shouldBe "0.35"
@@ -51,3 +51,5 @@ class LottoWinningTest : BehaviorSpec({
         }
     }
 })
+
+fun String.toNumbers() = this.split(",").map { it.toInt() }
